Fixed XCUIApplication resolve element issue
Summary:In Xcode 7.3 Apple changed how XCUIApplications are resolved.
In some cases trying to immediately resolve application will fail due to missing `_applicationQuery` property, which will lead to nilled `lastSnapshot` property that we depend on.
Calling `query` beforehand fixes that isses, because it sets `_applicationQuery`, if missing. Apple does the same thing when `lastSnapshot` is called ;].

Reviewed By: nqmtuan

Differential Revision: D3133298

fb-gh-sync-id: ef441c296e72b6ddf197022d04c271a1db1afcf0
fbshipit-source-id: ef441c296e72b6ddf197022d04c271a1db1afcf0