Added delay to /homescreen endpoint
Summary:
This is terrible workaround to the fact that pressButton:XCUIDeviceButtonHome is not a synchronous action.
On 9.2 some first queries  will trigger additional "go to home" event
So if we don't wait here it will be interpreted as double home button gesture and go to application switcher instead.
On 9.3 pressButton:XCUIDeviceButtonHome can be slightly delayed.
Causing waitUntilApplicationBoardIsVisible not to work properly in some edge cases e.g. like starting session right after this call, while being on home screen

Reviewed By: lawrencelomax

Differential Revision: D3223841

fb-gh-sync-id: 60e12f842b74a391fb05b6417a9a280848056e9b
fbshipit-source-id: 60e12f842b74a391fb05b6417a9a280848056e9b