Improved scroll to visible
Summary:
allow-large-files
When scrolling through long scrollviews XCUIElement we are pointing to might change. This is due to the fact that XCUIElement is not pointing to particular element, but it is a 'path' / 'query' for that element.
Because XCTWDA reports reused cells as well, during scroll that 'path' might start matching multiple elments and we will pick random one from them.

We can solve this by grabbing snapshot of element we are scrolling to and comparing visible elements to that snapshot. By doing that we effectivly we try to make strong match to required element.

Reviewed By: nqmtuan, lawrencelomax

Differential Revision: D3264101

fb-gh-sync-id: 61578bc333a47f8b5a03ce257060187d9c767a07
fbshipit-source-id: 61578bc333a47f8b5a03ce257060187d9c767a07