Improve deactivate app logic
Summary:
We used to fire home button and try to scroll right after that, assuming duration is 0.
It does not work well with app switching as XCTest often can get stuck while requesting elements.
Instead we should use `fb_goToHomescreenWithError` which handles that situation better and wait minimum 3 sec (We can really detect in which switch phase we are).

Reviewed By: lawrencelomax

Differential Revision: D4147432

fbshipit-source-id: 79de8be9d320bba9e149f9fc3503c08038251072