Improve xpath queries performance
Summary:
Current implementation takes up to 20 seconds to handle complicated XPath locators. With these changes the speed is improved to up to four times for some requests (5s vs 20s).
Changes made:
- KissXML framework is replaced with libxml2. This does not affect location performance very much. Only cases, where UI tree is really big, are affected. But now we have less external dependencies since libxml2 is built-in into iPhone SDK
- Made elements filtering procedure more flexible. This change has huge influence to the performance. I did some performance measurements and it seems, that the most time-costly operation here is snapshots comparison, because it is necessary to call

``` objc
[element resolve]
```

each time we need to get a snapshot. And this was done for every element in the UI. Now we don't do it and try to mach only elements, whose types can potentially match to snapshot types. Unfortunately, this does not make any difference if we are looking for _XCUIElementTypeAny_, but works pretty well (and fast
Closes https://github.com/facebook/WebDriverAgent/pull/308

Differential Revision: D4197241

Pulled By: marekcirkos

fbshipit-source-id: 2b77514984ae199943e8dd20a39b02e93b852e9d