Fix parsing of predicate string
Summary:
This was not able to get the actual suffix, since it modified the property name as it went. Instead it should pull the substrings from the original property name. Also `substringFromIndex` is inclusive, so we need to go from one past the position of the dot.
Closes https://github.com/facebook/WebDriverAgent/pull/456

Differential Revision: D4461673

Pulled By: marekcirkos

fbshipit-source-id: ee734ef672fa8630e61a0fb73cf74cc8d5dbdb82