Replace isVisible implementation
Summary:
Current isVisible implementation calls testmanagerd and at some conditions holds for over 60s. Since isVisible is pretty common call at times this can be really problematic.

This implementation is based on frame intersection and hitpoint checking.
In performance of both approaches is firmly similar, correctness looks the same as well.

However, hitpoint does make connection to testmanagerd as well so it may have same issues as isVisible querry.
Because isVisible was only call that caused the problem we have a good chance of fixing it, so let's try.

Reviewed By: lawrencelomax

Differential Revision: D4762527

fbshipit-source-id: 49f3b140a923877be5fea04a168f1b69d6084b01