Add capability to set typing speed
Summary:
Currently the maximum typing frequency is set at `60`. We have been seeing some cases, particularly when running on virtualized hardware, where clashes in the gestures occur and the operation fails (in fact the WebDriverAgent process fails altogether). Slowing down the frequency seems to be of great help.

So adding it as a desired capability, `maxTypingFrequency`, which can be sent in at session initialization time.
Closes https://github.com/facebook/WebDriverAgent/pull/529

Differential Revision: D4818642

Pulled By: marekcirkos

fbshipit-source-id: 0116e6ee84837875257ba4aa14c6dd60437b7045