Support clicking on a specific alert button
Summary:
Currently, the WebDriverAgent allows you to either accept or dismiss an alert (a leftover from HTML, where JavaScript alerts only defined two options).

The logic for deciding on which button to click when accepting or dismissing an alert is based on the location of the button - the web driver agent will click on either the first or the last button in the alert.

This PR attempts to make the webdriveragent a bit more flexible, by adding the following routes:

- `GET /alert/buttons`: this route returns the labels (`wdLabel`, same value as returned by `/element/:elementId/text`) of all buttons shown in the current alert
- `POST /alert/click/:label`: this route will try to close the alert by clicking on the button with label `:label`.

All feedback is welcome.
Closes https://github.com/facebook/WebDriverAgent/pull/410

Differential Revision: D4875461

Pulled By: marekcirkos

fbshipit-source-id: 002979f653f0b1a445039faf8f59117e93f12757