Make the offset configurable
Summary:
This PR adds an optional _offset_ attribute to _/wda/pickerwheel/:uuid/select_ API call. This is done, because the default offset, which was previously equal to 0.2 may not work as expected for some picker wheel implementations and make the wheel to switch more than one value at once. Now this behaviour is configurable.
Closes https://github.com/facebook/WebDriverAgent/pull/549

Differential Revision: D4921302

Pulled By: marekcirkos

fbshipit-source-id: 841495cb168ff63fe2fcb24e1ae74804cf1d4b75