Use dedicated screenshoting method for iOS 11
Summary:
The screenshot implementation based on XCAXClient_iOS is kind'a hacky and sometimes unstable (so we need to retry several times on Appium side or fall back to simctl in case of Simulator). And since iOS11 Apple finally added native support for making screenshots then why not to use it?
Closes https://github.com/facebook/WebDriverAgent/pull/660

Differential Revision: D5478308

Pulled By: marekcirkos

fbshipit-source-id: fa2ac247a5535c8468bb40a4041a0051407b342b