Use backspace-delete instead of just backspace for clear
Summary:
We have had issues where the cursor begins in the middle of the text, so just entering a backspace never finishes.

Here instead we enter a backspace (`\u0008`, moving to the left of the cursor) and a delete (`\u007F`, moving to the right of the cursor) together, so the string is deleted in both directions.

The weird machinations is to combat the error "universal character name refers to a control character" which came from trying to enter the unicode sequence directly, take [from here](https://stackoverflow.com/a/17155938).
Closes https://github.com/facebook/WebDriverAgent/pull/666

Differential Revision: D5571070

Pulled By: marekcirkos

fbshipit-source-id: 8e24bd0c76ca3035ef6dcd5c4ea1620cf378536e