Don't check if element is obstructed by alert while fetching.
Summary: This causes us to query for the alert, which goes through looking for sheets in springboard. That takes roughly 0.05 seconds each time and makes execution really slow for complicated actions. Note this is a behavioural change in the case element is actually obstructed, so the error might message for that might move somewhere else in the stack.

Reviewed By: lawrencelomax

Differential Revision: D5823287

fbshipit-source-id: 60c7adb6ef4a786d92bc95746784b06283bcf3d2