Fix navigate to homescreen for newer iOS versions
Summary:
On iOS 11 the Dock's `isVisible` property is often reported as false, even when the phone is on its home screen (and visual inspection show that the dock is visible and contain icons). In this situation the icons in the dock are all also reported as not visible.

Test that "Home screen icons" is visible instead, but keep using "Dock" on iOS 10.x and below ("Dock" work on those and "Home screen icons" is not present).

Reviewed By: antiarchit

Differential Revision: D5908395

fbshipit-source-id: 3a6740f13d6d8e4026150f6ace64881fc8812a1d