Add a possibility to use containingPredicate: selector in class chain queries
Summary:
For now class chain already implements all the available XCUIElementQuery selectors, except of the _containing..._ one. This PR adds support of containingPredicate: selector (which is also a superset of containingType:identifier: one), so now the implementation is full and provides API clients with complete set of selectors, which are supported natively by XCTest.
Closes https://github.com/facebook/WebDriverAgent/pull/707

Differential Revision: D6207919

Pulled By: marekcirkos

fbshipit-source-id: 9d0dd5ce06ff403916395338d6133211228e082b