Add one more source type
Summary:
There are many complains about _/source_ endpoint is slow, especially on iOS11. This is something, that cannot be avoided, but instead one can use the standard XCUIElement's _debugDescription_ property to see the basic hierarchy as Apple itself recommends. This endpoint is fast and might be enough for clients to just quickly check the current state of the UI tree.
Closes https://github.com/facebook/WebDriverAgent/pull/731

Differential Revision: D6207935

Pulled By: marekcirkos

fbshipit-source-id: 092f668b45b8975772216856e74d48337590048b