Use the native 'activate' method for application activation
Summary:
Since Xcode9 SDK Apple has added support for multiple applications in scope of a single session. This PR adds _activate_ endpoint to XCUIApplication instances and simplifies/makes more stable the implementation of _deactivateApp_ feature. This is going to be used as the base for the further PR, which adds multiple applications support to WDA.
Closes https://github.com/facebook/WebDriverAgent/pull/745

Differential Revision: D6208507

Pulled By: marekcirkos

fbshipit-source-id: 8c0d8f07264c961f85bb07d4f88d105b5c6e1e33