A new word in visibility detection
Summary:
This PR speeds up visibility detection in general, fixes it for landscape mode and also fixes scrolling to element for Xcode9+
Previously visibility detection was based on _hittable_ property value, but this property does not work for static elements, like an image in a cell, since hittablity for such elements is defined by gesture recogniser. Now we use another approach: we detect the visibility of the particular element based on the intersection between its own frame and all ancestor frames. This gives better results, since even if the element is not hittable the viability value is calculated properly for it:
![before](https://user-images.githubusercontent.com/7767781/32890265-5340e8ce-cace-11e7-9aae-26ebbcac73b8.png)
![now](https://user-images.githubusercontent.com/7767781/32890270-567eacb0-cace-11e7-9236-9da607f8325f.png)

As a result, the issues like https://github.com/facebook/WebDriverAgent/issues/746 should be addresses after the PR is merged and the overall visibility detection speed should be better.
Closes https://github.com/facebook/WebDriverAgent/pull/794

Reviewed By: marekcirkos

Differential Revision: D6457897

Pulled By: antiarchit

fbshipit-source-id: c4dff06c9cae7e53806377e4b5ff5009be0eea9b