Explicitly weak link XCTAutomationSupport.framework
Summary:
More and more classes are moved from public `XCTest` framework to private `XCTAutomationSupport` one.
It is not big of an issue, however in order to be able to compile WDA with new XCode we need to either:
- explicitly link both frameworks (this diff)
- or remove direct class allocations

Direct linking would remove requirement for future refactoring that might be necessary with potentially other moved classes so going for this approach.

Reviewed By: antiarchit

Differential Revision: D9081652

fbshipit-source-id: 94a8162eee491ced175a56f4381b2bba9ce31c16